fix: @skyqrose feedback- metadata channel gates other socket use
